+++ Bookstore/src/view/ConfirmEvent.java
@@ -4,24 +4,24 @@ import java.util.ArrayList;
 import java.util.EventObject;
 
 public class ConfirmEvent extends EventObject {
-	
+
 	private static final long serialVersionUID = 216887595747277843L;
-	private ArrayList<Integer> orders;
-	
+	private ArrayList<Long> orders;
+
 	public ConfirmEvent(Object source) {
 		super(source);
 	}
-	
-	public ConfirmEvent(Object source, ArrayList<Integer> orders) {
+
+	public ConfirmEvent(Object source, ArrayList<Long> orders) {
 		super(source);
 		this.orders = orders;
 	}
-	
-	ArrayList<Integer> getOrders() {
+
+	ArrayList<Long> getOrders() {
 		return orders;
 	}
-	
-	void setOrders(ArrayList<Integer> orders) {
+
+	void setOrders(ArrayList<Long> orders) {
 		this.orders = orders;
 	}
